refactor: split states/behaviours to avoid circular import
