--- packages/valory/skills/mech_interact_abci/behaviours.py
@@ -1,664 +1,0 @@
-# -*- coding: utf-8 -*-
-# ------------------------------------------------------------------------------
-#
-#   Copyright 2023 Valory AG
-#
-#   Licensed under the Apache License, Version 2.0 (the "License");
-#   you may not use this file except in compliance with the License.
-#   You may obtain a copy of the License at
-#
-#       http://www.apache.org/licenses/LICENSE-2.0
-#
-#   Unless required by applicable law or agreed to in writing, software
-#   distributed under the License is distributed on an "AS IS" BASIS,
-#   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-#   See the License for the specific language governing permissions and
-#   limitations under the License.
-#
-# ------------------------------------------------------------------------------
-
-"""This package contains round behaviours of MechInteractAbciApp."""
-
-import json
-from abc import ABC
-from dataclasses import asdict, is_dataclass
-from datetime import datetime, timedelta
-from pathlib import Path
-from tempfile import mkdtemp
-from typing import Any, Callable, Dict, Generator, List, Optional, Set, Type, cast
-
-import multibase
-import multicodec
-from aea.configurations.data_types import PublicId
-from aea.helpers.cid import to_v1
-from hexbytes import HexBytes
-from web3.constants import ADDRESS_ZERO
-
-from packages.valory.contracts.gnosis_safe.contract import (
-    GnosisSafeContract,
-    SafeOperation,
-)
-from packages.valory.contracts.mech.contract import Mech
-from packages.valory.contracts.multisend.contract import MultiSendContract
-from packages.valory.protocols.contract_api import ContractApiMessage
-from packages.valory.skills.abstract_round_abci.base import BaseTxPayload, get_name
-from packages.valory.skills.abstract_round_abci.behaviour_utils import (
-    BaseBehaviour,
-    TimeoutException,
-)
-from packages.valory.skills.abstract_round_abci.behaviours import AbstractRoundBehaviour
-from packages.valory.skills.abstract_round_abci.io_.store import SupportedFiletype
-from packages.valory.skills.mech_interact_abci.models import (
-    MechInteractionResponse,
-    MechMetadata,
-    MechParams,
-    MechRequest,
-    MechResponseSpecs,
-    MultisendBatch,
-)
-from packages.valory.skills.mech_interact_abci.payloads import (
-    MechRequestPayload,
-    MechResponsePayload,
-)
-from packages.valory.skills.mech_interact_abci.rounds import (
-    MechInteractAbciApp,
-    MechRequestRound,
-    MechResponseRound,
-    SynchronizedData,
-)
-from packages.valory.skills.transaction_settlement_abci.payload_tools import (
-    hash_payload_to_hex,
-)
-from packages.valory.skills.transaction_settlement_abci.rounds import TX_HASH_LENGTH
-
-
-WaitableConditionType = Generator[None, None, bool]
-
-METADATA_FILENAME = "metadata.json"
-V1_HEX_PREFIX = "f01"
-Ox = "0x"
-IPFS_HASH_PREFIX = f"{V1_HEX_PREFIX}701220"
-
-# setting the safe gas to 0 means that all available gas will be used
-# which is what we want in most cases
-# more info here: https://safe-docs.dev.gnosisdev.com/safe/docs/contracts_tx_execution/
-SAFE_GAS = 0
-
-
-class DataclassEncoder(json.JSONEncoder):
-    """A custom JSON encoder for dataclasses."""
-
-    def default(self, o: Any) -> Any:
-        """The default JSON encoder."""
-        if is_dataclass(o):
-            return asdict(o)
-        return super().default(o)
-
-
-class MechInteractBaseBehaviour(BaseBehaviour, ABC):
-    """Represents the base class for the mech interaction FSM behaviour."""
-
-    def __init__(self, **kwargs: Any) -> None:
-        """Initialize the bet placement behaviour."""
-        super().__init__(**kwargs)
-        self.multisend_batches: List[MultisendBatch] = []
-        self.multisend_data = b""
-        self._safe_tx_hash = ""
-
-    @property
-    def synchronized_data(self) -> SynchronizedData:
-        """Return the synchronized data."""
-        return cast(SynchronizedData, super().synchronized_data)
-
-    @property
-    def params(self) -> MechParams:
-        """Return the params."""
-        return cast(MechParams, self.context.params)
-
-    def default_error(
-        self, contract_id: str, contract_callable: str, response_msg: ContractApiMessage
-    ) -> None:
-        """Return a default contract interaction error message."""
-        self.context.logger.error(
-            f"Could not successfully interact with the {contract_id} contract "
-            f"using {contract_callable!r}: {response_msg}"
-        )
-
-    def contract_interaction_error(
-        self, contract_id: str, contract_callable: str, response_msg: ContractApiMessage
-    ) -> None:
-        """Return a contract interaction error message."""
-        # contracts can only return one message, i.e., multiple levels cannot exist.
-        for level in ("info", "warning", "error"):
-            msg = response_msg.raw_transaction.body.get(level, None)
-            logger = getattr(self.context.logger, level)
-            if msg is not None:
-                logger(msg)
-                return
-
-        self.default_error(contract_id, contract_callable, response_msg)
-
-    def contract_interact(
-        self,
-        performative: ContractApiMessage.Performative,
-        contract_address: str,
-        contract_public_id: PublicId,
-        contract_callable: str,
-        data_key: str,
-        placeholder: str,
-        **kwargs: Any,
-    ) -> WaitableConditionType:
-        """Interact with a contract."""
-        contract_id = str(contract_public_id)
-        response_msg = yield from self.get_contract_api_response(
-            performative,
-            contract_address,
-            contract_id,
-            contract_callable,
-            **kwargs,
-        )
-        if response_msg.performative != ContractApiMessage.Performative.RAW_TRANSACTION:
-            self.default_error(contract_id, contract_callable, response_msg)
-            return False
-
-        data = response_msg.raw_transaction.body.get(data_key, None)
-        if data is None:
-            self.contract_interaction_error(
-                contract_id, contract_callable, response_msg
-            )
-            return False
-
-        setattr(self, placeholder, data)
-        return True
-
-    def _mech_contract_interact(
-        self, contract_callable: str, data_key: str, placeholder: str, **kwargs: Any
-    ) -> WaitableConditionType:
-        """Interact with the mech contract."""
-        status = yield from self.contract_interact(
-            performative=ContractApiMessage.Performative.GET_RAW_TRANSACTION,  # type: ignore
-            contract_address=self.params.mech_agent_address,
-            contract_public_id=Mech.contract_id,
-            contract_callable=contract_callable,
-            data_key=data_key,
-            placeholder=placeholder,
-            **kwargs,
-        )
-        return status
-
-    def wait_for_condition_with_sleep(
-        self,
-        condition_gen: Callable[[], WaitableConditionType],
-        timeout: Optional[float] = None,
-    ) -> Generator:
-        """Wait for a condition to happen and sleep in-between checks.
-
-        This is a modified version of the base `wait_for_condition` method which:
-            1. accepts a generator that creates the condition instead of a callable
-            2. sleeps in-between checks
-
-        :param condition_gen: a generator of the condition to wait for
-        :param timeout: the maximum amount of time to wait
-        :yield: None
-        """
-
-        deadline = (
-            datetime.now() + timedelta(0, timeout)
-            if timeout is not None
-            else datetime.max
-        )
-
-        while True:
-            condition_satisfied = yield from condition_gen()
-            if condition_satisfied:
-                break
-            if timeout is not None and datetime.now() > deadline:
-                raise TimeoutException()
-            self.context.logger.info(f"Retrying in {self.params.sleep_time} seconds.")
-            yield from self.sleep(self.params.sleep_time)
-
-    def finish_behaviour(self, payload: BaseTxPayload) -> Generator:
-        """Finish the behaviour."""
-        with self.context.benchmark_tool.measure(self.behaviour_id).consensus():
-            yield from self.send_a2a_transaction(payload)
-            yield from self.wait_until_round_end()
-
-        self.set_done()
-
-
-class MechRequestBehaviour(MechInteractBaseBehaviour):
-    """A behaviour in which the agents prepare a tx to initiate a request to a mech."""
-
-    matching_round = MechRequestRound
-
-    def __init__(self, **kwargs: Any) -> None:
-        """Initialize Behaviour."""
-        super().__init__(**kwargs)
-        self._v1_hex_truncated: str = ""
-        self._request_data: bytes = b""
-        self._price: int = 0
-        self._mech_requests: List[MechMetadata] = []
-        self._pending_responses: List[MechInteractionResponse] = []
-
-    @property
-    def metadata_filepath(self) -> str:
-        """Get the filepath to the metadata."""
-        return str(Path(mkdtemp()) / METADATA_FILENAME)
-
-    @property
-    def request_data(self) -> bytes:
-        """Get the request data."""
-        return self._request_data
-
-    @request_data.setter
-    def request_data(self, data: bytes) -> None:
-        """Set the request data."""
-        self._request_data = data
-
-    @property
-    def price(self) -> int:
-        """Get the price."""
-        return self._price
-
-    @price.setter
-    def price(self, price: int) -> None:
-        """Set the price."""
-        self._price = price
-
-    @property
-    def safe_tx_hash(self) -> str:
-        """Get the safe_tx_hash."""
-        return self._safe_tx_hash
-
-    @safe_tx_hash.setter
-    def safe_tx_hash(self, safe_hash: str) -> None:
-        """Set the safe_tx_hash."""
-        length = len(safe_hash)
-        if length != TX_HASH_LENGTH:
-            raise ValueError(
-                f"Incorrect length {length} != {TX_HASH_LENGTH} detected "
-                f"when trying to assign a safe transaction hash: {safe_hash}"
-            )
-        self._safe_tx_hash = safe_hash[2:]
-
-    @property
-    def multi_send_txs(self) -> List[dict]:
-        """Get the multisend transactions as a list of dictionaries."""
-        return [asdict(batch) for batch in self.multisend_batches]
-
-    @property
-    def txs_value(self) -> int:
-        """Get the total value of the transactions."""
-        return sum(batch.value for batch in self.multisend_batches)
-
-    @property
-    def tx_hex(self) -> Optional[str]:
-        """Serialize the safe tx to a hex string."""
-        if self.safe_tx_hash == "":
-            raise ValueError(
-                "Cannot prepare a multisend transaction without a safe transaction hash."
-            )
-        return hash_payload_to_hex(
-            self.safe_tx_hash,
-            self.txs_value,
-            SAFE_GAS,
-            self.params.multisend_address,
-            self.multisend_data,
-            SafeOperation.DELEGATE_CALL.value,
-        )
-
-    def _build_multisend_data(
-        self,
-    ) -> WaitableConditionType:
-        """Get the multisend tx."""
-        response_msg = yield from self.get_contract_api_response(
-            performative=ContractApiMessage.Performative.GET_RAW_TRANSACTION,  # type: ignore
-            contract_address=self.params.multisend_address,
-            contract_id=str(MultiSendContract.contract_id),
-            contract_callable="get_tx_data",
-            multi_send_txs=self.multi_send_txs,
-        )
-        expected_performative = ContractApiMessage.Performative.RAW_TRANSACTION
-        if response_msg.performative != expected_performative:
-            self.context.logger.error(
-                f"Couldn't compile the multisend tx. "
-                f"Expected response performative {expected_performative.value}, "  # type: ignore
-                f"received {response_msg.performative.value}: {response_msg}"
-            )
-            return False
-
-        multisend_data_str = response_msg.raw_transaction.body.get("data", None)
-        if multisend_data_str is None:
-            self.context.logger.error(
-                f"Something went wrong while trying to prepare the multisend data: {response_msg}"
-            )
-            return False
-
-        # strip "0x" from the response
-        multisend_data_str = str(response_msg.raw_transaction.body["data"])[2:]
-        self.multisend_data = bytes.fromhex(multisend_data_str)
-        return True
-
-    def _build_multisend_safe_tx_hash(self) -> WaitableConditionType:
-        """Prepares and returns the safe tx hash for a multisend tx."""
-        response_msg = yield from self.get_contract_api_response(
-            performative=ContractApiMessage.Performative.GET_STATE,  # type: ignore
-            contract_address=self.synchronized_data.safe_contract_address,
-            contract_id=str(GnosisSafeContract.contract_id),
-            contract_callable="get_raw_safe_transaction_hash",
-            to_address=self.params.multisend_address,
-            value=self.txs_value,
-            data=self.multisend_data,
-            safe_tx_gas=SAFE_GAS,
-            operation=SafeOperation.DELEGATE_CALL.value,
-        )
-
-        if response_msg.performative != ContractApiMessage.Performative.STATE:
-            self.context.logger.error(
-                "Couldn't get safe tx hash. Expected response performative "
-                f"{ContractApiMessage.Performative.STATE.value}, "  # type: ignore
-                f"received {response_msg.performative.value}: {response_msg}."
-            )
-            return False
-
-        tx_hash = response_msg.state.body.get("tx_hash", None)
-        if tx_hash is None or len(tx_hash) != TX_HASH_LENGTH:
-            self.context.logger.error(
-                "Something went wrong while trying to get the buy transaction's hash. "
-                f"Invalid hash {tx_hash!r} was returned."
-            )
-            return False
-
-        self.safe_tx_hash = tx_hash
-        return True
-
-    def setup(self) -> None:
-        """Set up the `MechRequest` behaviour."""
-        self._mech_requests = self.synchronized_data.mech_requests
-
-    def _send_metadata_to_ipfs(
-        self,
-    ) -> WaitableConditionType:
-        """Send Mech metadata to IPFS."""
-        metadata = self._mech_requests.pop()
-        metadata_hash = yield from self.send_to_ipfs(
-            self.metadata_filepath, asdict(metadata), filetype=SupportedFiletype.JSON
-        )
-        if metadata_hash is None:
-            return False
-
-        v1_file_hash = to_v1(metadata_hash)
-        cid_bytes = cast(bytes, multibase.decode(v1_file_hash))
-        multihash_bytes = multicodec.remove_prefix(cid_bytes)
-        v1_file_hash_hex = V1_HEX_PREFIX + multihash_bytes.hex()
-        ipfs_link = self.params.ipfs_address + v1_file_hash_hex
-        self.context.logger.info(f"Prompt uploaded: {ipfs_link}")
-        mech_request_data = v1_file_hash_hex[9:]
-        pending_response = MechInteractionResponse(
-            nonce=metadata.nonce, data=mech_request_data
-        )
-        self._v1_hex_truncated = Ox + mech_request_data
-        self._pending_responses.append(pending_response)
-        return True
-
-    def _build_request_data(self) -> WaitableConditionType:
-        """Get the request tx data encoded."""
-        status = yield from self._mech_contract_interact(
-            "get_request_data",
-            "data",
-            get_name(MechRequestBehaviour.request_data),
-            request_data=self._v1_hex_truncated,
-        )
-
-        if status:
-            batch = MultisendBatch(
-                to=self.params.mech_agent_address,
-                data=HexBytes(self.request_data),
-                value=self.price,
-            )
-            self.multisend_batches.append(batch)
-
-        return status
-
-    def _get_price(self) -> WaitableConditionType:
-        """Get the price of the mech request."""
-        result = yield from self._mech_contract_interact(
-            "get_price", "price", get_name(MechRequestBehaviour.price)
-        )
-        return result
-
-    def _prepare_safe_tx(self) -> Generator:
-        """Prepare a multisend safe tx for sending requests to a mech and return the hex for the tx settlement skill."""
-        n_iters = min(self.params.multisend_batch_size, len(self._mech_requests))
-        steps = (self._get_price,)
-        steps += (self._send_metadata_to_ipfs, self._build_request_data) * n_iters
-        steps += (self._build_multisend_data, self._build_multisend_safe_tx_hash)
-
-        for step in steps:
-            yield from self.wait_for_condition_with_sleep(step)
-
-    def async_act(self) -> Generator:
-        """Do the action."""
-
-        with self.context.benchmark_tool.measure(self.behaviour_id).local():
-            yield from self._prepare_safe_tx()
-            serialized_data = (
-                json.dumps(data, cls=DataclassEncoder)
-                for data in (self._mech_requests, self._pending_responses)
-            )
-            payload = MechRequestPayload(
-                self.context.agent_address, self.tx_hex, self.price, *serialized_data
-            )
-        yield from self.finish_behaviour(payload)
-
-
-class MechResponseBehaviour(MechInteractBaseBehaviour):
-    """A behaviour in which the agents receive the Mech's responses."""
-
-    matching_round = MechResponseRound
-
-    def __init__(self, **kwargs: Any) -> None:
-        """Initialize Behaviour."""
-        super().__init__(**kwargs)
-        self._from_block: int = 0
-        self._requests: List[MechRequest] = []
-        self._response_hex: str = ""
-        self._mech_responses: List[MechInteractionResponse] = []
-        self._current_mech_response: MechInteractionResponse = MechInteractionResponse(
-            error="The mech's response has not been set!"
-        )
-
-    @property
-    def from_block(self) -> int:
-        """Get the block number in which the request to the mech was settled."""
-        return self._from_block
-
-    @from_block.setter
-    def from_block(self, from_block: int) -> None:
-        """Set the block number in which the request to the mech was settled."""
-        self._from_block = from_block
-
-    @property
-    def requests(self) -> List[MechRequest]:
-        """Get the requests."""
-        return self._requests
-
-    @requests.setter
-    def requests(self, requests: List[Dict[str, str]]) -> None:
-        """Set the requests."""
-        self._requests = [MechRequest(**request) for request in requests]
-
-    @property
-    def response_hex(self) -> str:
-        """Get the hash of the response data."""
-        return self._response_hex
-
-    @response_hex.setter
-    def response_hex(self, response_hash: bytes) -> None:
-        """Set the hash of the response data."""
-        try:
-            self._response_hex = response_hash.hex()
-        except AttributeError:
-            msg = f"Response hash {response_hash!r} is not valid hex bytes!"
-            self.context.logger.error(msg)
-
-    @property
-    def mech_response_api(self) -> MechResponseSpecs:
-        """Get the mech response api specs."""
-        return self.context.mech_response
-
-    @property
-    def serialized_responses(self) -> str:
-        """Get the Mech's responses serialized."""
-        return json.dumps(self._mech_responses, cls=DataclassEncoder)
-
-    def setup(self) -> None:
-        """Set up the `MechResponse` behaviour."""
-        self._mech_responses = self.synchronized_data.mech_responses
-
-    def set_mech_response_specs(self, request_id: int) -> None:
-        """Set the mech's response specs."""
-        full_ipfs_hash = IPFS_HASH_PREFIX + self.response_hex
-        ipfs_link = self.params.ipfs_address + full_ipfs_hash + f"/{request_id}"
-        # The url must be dynamically generated as it depends on the ipfs hash
-        self.mech_response_api.__dict__["_frozen"] = False
-        self.mech_response_api.url = ipfs_link
-        self.mech_response_api.__dict__["_frozen"] = True
-
-    def _get_block_number(self) -> WaitableConditionType:
-        """Get the block number in which the request to the mech was settled."""
-        result = yield from self.contract_interact(
-            performative=ContractApiMessage.Performative.GET_RAW_TRANSACTION,  # type: ignore
-            # we do not need the address to get the block number, but the base method does
-            contract_address=ADDRESS_ZERO,
-            contract_public_id=Mech.contract_id,
-            contract_callable="get_block_number",
-            data_key="number",
-            placeholder=get_name(MechResponseBehaviour.from_block),
-            tx_hash=self.synchronized_data.final_tx_hash,
-        )
-
-        return result
-
-    def _process_request_event(self) -> WaitableConditionType:
-        """Process the request event."""
-        result = yield from self._mech_contract_interact(
-            contract_callable="process_request_event",
-            data_key="results",
-            placeholder=get_name(MechResponseBehaviour.requests),
-            tx_hash=self.synchronized_data.final_tx_hash,
-            expected_logs=len(self._mech_responses),
-        )
-        return result
-
-    def _get_response_hash(self) -> WaitableConditionType:
-        """Get the hash of the response data."""
-        request_id = self._current_mech_response.requestId
-        self.context.logger.info(
-            f"Filtering the mech's events from block {self.from_block} "
-            f"for a response to our request with id {request_id!r}."
-        )
-        result = yield from self._mech_contract_interact(
-            contract_callable="get_response",
-            data_key="data",
-            placeholder=get_name(MechResponseBehaviour.response_hex),
-            request_id=request_id,
-            from_block=self.from_block,
-        )
-
-        if result:
-            self.set_mech_response_specs(request_id)
-
-        return result
-
-    def _handle_response(
-        self,
-        res: Optional[str],
-    ) -> Optional[Any]:
-        """Handle the response from the IPFS.
-
-        :param res: the response to handle.
-        :return: the response's result, using the given keys. `None` if response is `None` (has failed).
-        """
-        if res is None:
-            msg = f"Could not get the mech's response from {self.mech_response_api.api_id}"
-            self.context.logger.error(msg)
-            self.mech_response_api.increment_retries()
-            return None
-
-        self.context.logger.info(f"Retrieved the mech's response: {res}.")
-        self.mech_response_api.reset_retries()
-        return res
-
-    def _get_response(self) -> WaitableConditionType:
-        """Get the response data from IPFS."""
-        specs = self.mech_response_api.get_spec()
-        res_raw = yield from self.get_http_response(**specs)
-        res = self.mech_response_api.process_response(res_raw)
-        res = self._handle_response(res)
-
-        if self.mech_response_api.is_retries_exceeded():
-            self._current_mech_response.retries_exceeded()
-            return True
-
-        if res is None:
-            return False
-
-        try:
-            self._current_mech_response.result = res
-        except (ValueError, TypeError):
-            self._current_mech_response.incorrect_format(res)
-
-        return True
-
-    def _set_current_response(self, request: MechRequest) -> None:
-        """Set the current Mech response."""
-        for pending_response in self._mech_responses:
-            if pending_response.data == request.data:
-                pending_response.requestId = request.requestId
-                self._current_mech_response = pending_response
-                break
-
-    def _process_responses(
-        self,
-    ) -> Generator:
-        """Get the response."""
-        for step in (
-            self._get_block_number,
-            self._process_request_event,
-        ):
-            yield from self.wait_for_condition_with_sleep(step)
-
-        for request in self.requests:
-            self._set_current_response(request)
-
-            for step in (self._get_response_hash, self._get_response):
-                yield from self.wait_for_condition_with_sleep(step)
-
-            self.context.logger.info(
-                f"Response has been received:\n{self._current_mech_response}"
-            )
-            if self._current_mech_response.result is None:
-                self.context.logger.error(
-                    f"There was an error in the mech's response: {self._current_mech_response.error}"
-                )
-
-    def async_act(self) -> Generator:
-        """Do the action."""
-
-        with self.context.benchmark_tool.measure(self.behaviour_id).local():
-            yield from self._process_responses()
-            payload = MechResponsePayload(
-                self.context.agent_address,
-                self.serialized_responses,
-            )
-
-        yield from self.finish_behaviour(payload)
-
-
-class MechInteractRoundBehaviour(AbstractRoundBehaviour):
-    """MechInteractRoundBehaviour"""
-
-    initial_behaviour_cls = MechRequestBehaviour
-    abci_app_cls = MechInteractAbciApp  # type: ignore
-    behaviours: Set[Type[BaseBehaviour]] = {MechRequestBehaviour, MechResponseBehaviour}
