+++ packages/valory/skills/mech_interact_abci/behaviours/base.py
@@ -1,0 +1,182 @@
+# -*- coding: utf-8 -*-
+# ------------------------------------------------------------------------------
+#
+#   Copyright 2023 Valory AG
+#
+#   Licensed under the Apache License, Version 2.0 (the "License");
+#   you may not use this file except in compliance with the License.
+#   You may obtain a copy of the License at
+#
+#       http://www.apache.org/licenses/LICENSE-2.0
+#
+#   Unless required by applicable law or agreed to in writing, software
+#   distributed under the License is distributed on an "AS IS" BASIS,
+#   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#   See the License for the specific language governing permissions and
+#   limitations under the License.
+#
+# ------------------------------------------------------------------------------
+
+"""This module contains the base behaviour for the mech interact abci skill."""
+
+import json
+from abc import ABC
+from dataclasses import asdict, is_dataclass
+from datetime import datetime, timedelta
+from typing import Any, Callable, Generator, List, Optional, cast
+
+from aea.configurations.data_types import PublicId
+
+from packages.valory.contracts.mech.contract import Mech
+from packages.valory.protocols.contract_api import ContractApiMessage
+from packages.valory.skills.abstract_round_abci.base import BaseTxPayload
+from packages.valory.skills.abstract_round_abci.behaviour_utils import (
+    BaseBehaviour,
+    TimeoutException,
+)
+from packages.valory.skills.mech_interact_abci.models import MechParams, MultisendBatch
+from packages.valory.skills.mech_interact_abci.states.base import SynchronizedData
+
+
+WaitableConditionType = Generator[None, None, bool]
+
+
+class MechInteractBaseBehaviour(BaseBehaviour, ABC):
+    """Represents the base class for the mech interaction FSM behaviour."""
+
+    def __init__(self, **kwargs: Any) -> None:
+        """Initialize the bet placement behaviour."""
+        super().__init__(**kwargs)
+        self.multisend_batches: List[MultisendBatch] = []
+        self.multisend_data = b""
+        self._safe_tx_hash = ""
+
+    @property
+    def synchronized_data(self) -> SynchronizedData:
+        """Return the synchronized data."""
+        return cast(SynchronizedData, super().synchronized_data)
+
+    @property
+    def params(self) -> MechParams:
+        """Return the params."""
+        return cast(MechParams, self.context.params)
+
+    def default_error(
+        self, contract_id: str, contract_callable: str, response_msg: ContractApiMessage
+    ) -> None:
+        """Return a default contract interaction error message."""
+        self.context.logger.error(
+            f"Could not successfully interact with the {contract_id} contract "
+            f"using {contract_callable!r}: {response_msg}"
+        )
+
+    def contract_interaction_error(
+        self, contract_id: str, contract_callable: str, response_msg: ContractApiMessage
+    ) -> None:
+        """Return a contract interaction error message."""
+        # contracts can only return one message, i.e., multiple levels cannot exist.
+        for level in ("info", "warning", "error"):
+            msg = response_msg.raw_transaction.body.get(level, None)
+            logger = getattr(self.context.logger, level)
+            if msg is not None:
+                logger(msg)
+                return
+
+        self.default_error(contract_id, contract_callable, response_msg)
+
+    def contract_interact(
+        self,
+        performative: ContractApiMessage.Performative,
+        contract_address: str,
+        contract_public_id: PublicId,
+        contract_callable: str,
+        data_key: str,
+        placeholder: str,
+        **kwargs: Any,
+    ) -> WaitableConditionType:
+        """Interact with a contract."""
+        contract_id = str(contract_public_id)
+        response_msg = yield from self.get_contract_api_response(
+            performative,
+            contract_address,
+            contract_id,
+            contract_callable,
+            **kwargs,
+        )
+        if response_msg.performative != ContractApiMessage.Performative.RAW_TRANSACTION:
+            self.default_error(contract_id, contract_callable, response_msg)
+            return False
+
+        data = response_msg.raw_transaction.body.get(data_key, None)
+        if data is None:
+            self.contract_interaction_error(
+                contract_id, contract_callable, response_msg
+            )
+            return False
+
+        setattr(self, placeholder, data)
+        return True
+
+    def _mech_contract_interact(
+        self, contract_callable: str, data_key: str, placeholder: str, **kwargs: Any
+    ) -> WaitableConditionType:
+        """Interact with the mech contract."""
+        status = yield from self.contract_interact(
+            performative=ContractApiMessage.Performative.GET_RAW_TRANSACTION,  # type: ignore
+            contract_address=self.params.mech_agent_address,
+            contract_public_id=Mech.contract_id,
+            contract_callable=contract_callable,
+            data_key=data_key,
+            placeholder=placeholder,
+            **kwargs,
+        )
+        return status
+
+    def wait_for_condition_with_sleep(
+        self,
+        condition_gen: Callable[[], WaitableConditionType],
+        timeout: Optional[float] = None,
+    ) -> Generator:
+        """Wait for a condition to happen and sleep in-between checks.
+
+        This is a modified version of the base `wait_for_condition` method which:
+            1. accepts a generator that creates the condition instead of a callable
+            2. sleeps in-between checks
+
+        :param condition_gen: a generator of the condition to wait for
+        :param timeout: the maximum amount of time to wait
+        :yield: None
+        """
+
+        deadline = (
+            datetime.now() + timedelta(0, timeout)
+            if timeout is not None
+            else datetime.max
+        )
+
+        while True:
+            condition_satisfied = yield from condition_gen()
+            if condition_satisfied:
+                break
+            if timeout is not None and datetime.now() > deadline:
+                raise TimeoutException()
+            self.context.logger.info(f"Retrying in {self.params.sleep_time} seconds.")
+            yield from self.sleep(self.params.sleep_time)
+
+    def finish_behaviour(self, payload: BaseTxPayload) -> Generator:
+        """Finish the behaviour."""
+        with self.context.benchmark_tool.measure(self.behaviour_id).consensus():
+            yield from self.send_a2a_transaction(payload)
+            yield from self.wait_until_round_end()
+
+        self.set_done()
+
+
+class DataclassEncoder(json.JSONEncoder):
+    """A custom JSON encoder for dataclasses."""
+
+    def default(self, o: Any) -> Any:
+        """The default JSON encoder."""
+        if is_dataclass(o):
+            return asdict(o)
+        return super().default(o)
