+++ packages/valory/skills/mech_interact_abci/behaviours/round_behaviour.py
@@ -1,0 +1,40 @@
+# -*- coding: utf-8 -*-
+# ------------------------------------------------------------------------------
+#
+#   Copyright 2023 Valory AG
+#
+#   Licensed under the Apache License, Version 2.0 (the "License");
+#   you may not use this file except in compliance with the License.
+#   You may obtain a copy of the License at
+#
+#       http://www.apache.org/licenses/LICENSE-2.0
+#
+#   Unless required by applicable law or agreed to in writing, software
+#   distributed under the License is distributed on an "AS IS" BASIS,
+#   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#   See the License for the specific language governing permissions and
+#   limitations under the License.
+#
+# ------------------------------------------------------------------------------
+
+"""This package contains the round behaviour of MechInteractAbciApp."""
+
+from typing import Set, Type
+
+from packages.valory.skills.abstract_round_abci.behaviour_utils import BaseBehaviour
+from packages.valory.skills.abstract_round_abci.behaviours import AbstractRoundBehaviour
+from packages.valory.skills.mech_interact_abci.behaviours.request import (
+    MechRequestBehaviour,
+)
+from packages.valory.skills.mech_interact_abci.behaviours.response import (
+    MechResponseBehaviour,
+)
+from packages.valory.skills.mech_interact_abci.rounds import MechInteractAbciApp
+
+
+class MechInteractRoundBehaviour(AbstractRoundBehaviour):
+    """MechInteractRoundBehaviour"""
+
+    initial_behaviour_cls = MechRequestBehaviour
+    abci_app_cls = MechInteractAbciApp  # type: ignore
+    behaviours: Set[Type[BaseBehaviour]] = {MechRequestBehaviour, MechResponseBehaviour}
