+++ packages/valory/skills/twitter_scoring_abci/rounds.py
@@ -35,7 +35,7 @@ from packages.valory.skills.abstract_round_abci.base import (
     EventToTimeout,
     get_name,
 )
-from packages.valory.skills.mech_interact_abci.models import (
+from packages.valory.skills.mech_interact_abci.states.base import (
     MechInteractionResponse,
     MechMetadata,
 )
